Add hooks for bundle installing "next" gem spec
we need a before-all hook to grab the previous lockfile state that we check
in an after-all hook. if there's changes, we no-op. otherwise, we build
a definition for "next" and pass it the `unlock` parameter from the
current definition. this should basically be empty but there's a few
key/values that it is best to pass through. finally, we fire off the
installer.
NOTE: this doesn't pass in the options parameters we get from the
command line