Bundle next on initial plugin install
This is sort of a bug with bundler plugins. When a plugin is installed
it won't run the registered hooks and requires another `bundle`. We can
fix this by extracting our "bundle next" into a method and calling that
at the time of load. Further, we can get better control over plugin load
by calling a method in `plugins.rb` instead of running code at the
global-level at file loadtime