Change clean subcommand spec
Later versions of bundler ie. 2.1.4 don't support using the `--help`
flag for plugin subcommands. This may be a bug or intended, either way
it's better to test an actual subcommand flag that is handled by our
code. We have a bit of duplication by using `--dry-run` in both specs,
but using `--force` would cause things to fail as it would actually
remove gems from the system. The duplication isn't all that bad since
we're testing for different parts, even tho the former is implicitly
covered by the latter.